check if user has u2f on superuser modal
